+++ CHANGES.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+1.0.1 (2022-06-02)
+------------------
+
+- bugfix for empty query and empty resultset in post handler
+- removing limit in post handler
+
 
 1.0.0 (2022-05-09)
 ------------------